% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAZooData.R
\name{DNAZooData}
\alias{DNAZooData}
\title{DNAZooData}
\arguments{
\item{species}{Any species processed by the DNA Zoo
(check https://www.dnazoo.org/assemblies) for a browser-based
explorer.}
}
\value{
\code{DNAZooData()} returns a HicFile object, which can then be
imported in memory using \code{HiCExperiment::import()}.
Metadata also points to a URL to directly fetch the genome assembly
corrected by the DNA Zoo consortium.
}
\description{
Fetches files from the DNAZoo data portal and caches them using
the BiocFileCache system.
}
\examples{
###################################
## Importing DNAZoo `.hic` files ##
###################################

head(DNAZooData())
hf <- DNAZooData(species = 'Anolis_carolinensis')
hf

}
