% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{DNAZooDataCache}
\alias{DNAZooDataCache}
\title{Manage cache / download files from the DNAZoo data portal}
\usage{
DNAZooDataCache(...)
}
\arguments{
\item{...}{Arguments passed to internal \code{.setDNAZooDataCache} function}
}
\value{
BiocFileCache object
}
\description{
Managing DNAZoo data downloads via the integrated
\code{BiocFileCache} system.
}
\examples{
bfc <- DNAZooDataCache()
bfc
BiocFileCache::bfcinfo(bfc)
}
